[Params][Trivial][Fork] Start shifting and refactoring code
